Limit user permissions

+++ portfolio/src/main/java/com/google/sps/servlets/DeleteCommentServlet.java
@@ -26,6 +26,8 @@ import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.EntityNotFoundException;
 import com.google.appengine.api.datastore.KeyFactory;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
 
 /* Servlet that adds a comment to the dataserve. */
 @WebServlet("/delete-comment")
@@ -34,11 +36,18 @@ public class DeleteCommentServlet extends HttpServlet {
   /* Removes the request comment from the datastore */
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    UserService userService = UserServiceFactory.getUserService();
+
     DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
     long id = Long.parseLong(request.getParameter("id"));
+    String currentUserEmail = userService.getCurrentUser().getEmail();
 
     try {
       Entity comment = datastore.get(KeyFactory.createKey("Comment", id));
+
+      // Don't allow if not an admin and not their comment
+      if (!userService.isUserAdmin() && comment.getProperty("email") != currentUserEmail) return; 
+
       datastore.delete(comment.getKey());
     } catch (EntityNotFoundException e) {
       return;
